Rolled back to have IdentifiableObject external to EventHandler.
