+++ Sources/Scenes/EntityMouseDragHandler.swift
@@ -16,7 +16,7 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-public protocol EntityMouseDragHandler : RenderableEntity {
+public protocol EntityMouseDragHandler : EventHandler, RenderableEntity {
     func onEntityMouseDrag(globalLocation:Point, movement:Point)
 }
