+++ Sources/Scenes/EntityMouseUpHandler.swift
@@ -16,7 +16,7 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-public protocol EntityMouseUpHandler : RenderableEntity {
+public protocol EntityMouseUpHandler : EventHandler, RenderableEntity {
     func onEntityMouseUp(globalLocation:Point)
 }
